ombres

+++ src/Point.h
@@ -1,17 +1,17 @@
 
 class Point {
 private:
-	double x;
-	double y;
-	double z;
+	float x;
+	float y;
+	float z;
 public:
-	Point(double, double, double);
-	double getX();
-	double getY();
-	double getZ();
-	void setX(double x);
-	void setY(double y);
-	void setZ(double Z);
+	Point(float, float, float);
+	float getX();
+	float getY();
+	float getZ();
+	void setX(float x);
+	void setY(float y);
+	void setZ(float Z);
 };
